+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to the Docker Language Server will be documented in this fil
 
 ### Fixed
 
+- Compose
+  - textDocument/documentSymbol
+    - prevent scalar values from showing up as a document symbol
+
+## [0.3.3] - 2025-04-09
+
+### Fixed
+
 - refactored the panic handler so that crashes from handling the JSON-RPC messages would no longer cause the language server to crash
 
 ## [0.3.2] - 2025-04-09
@@ -95,7 +103,8 @@ All notable changes to the Docker Language Server will be documented in this fil
   - textDocument/semanticTokens/full
     - provide syntax highlighting for Bake files
 
-[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.3.2...main
+[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.3.3...main
+[0.3.3]: https://github.com/docker/docker-language-server/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/docker/docker-language-server/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/docker/docker-language-server/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/docker/docker-language-server/compare/v0.2.0...v0.3.0